people module directory restructuring
