+++ lib/process-services/people/components/people-list/people-list.component.ts
@@ -18,7 +18,7 @@
 import { DataTableComponent } from '@alfresco/adf-core';
 import { DataColumnListComponent, UserProcessModel } from '@alfresco/adf-core';
 import { AfterContentInit, AfterViewInit, Component, ContentChild, EventEmitter, Input, Output, ViewChild } from '@angular/core';
-import { UserEventModel } from '../task-list/models/user-event.model';
+import { UserEventModel } from '../../../task-list/models/user-event.model';
 
 @Component({
     selector: 'adf-people-list',
