+++ lib/process-services/people/components/people-search/people-search.component.scss
