+++ lib/process-services/people/components/people/people.component.scss
@@ -51,13 +51,13 @@
         padding: 0px;
     }
 
-    
+
 
     adf-people-list adf-datatable {
         thead {
             display: none;
         }
-        
+
         .adf-data-table .adf-data-table-cell .cell-container {
             flex-direction: column;
             align-items: left;
