+++ lib/process-services/styles/_index.scss
@@ -2,8 +2,7 @@
 @import '../attachment/process-attachment-list.component';
 @import '../attachment/task-attachment-list.component';
 @import '../comments/comment-list.component';
-@import '../people/people-search.component';
-@import '../people/people.component';
+@import '../people/people.module';
 @import '../task-list/components/start-task.component';
 @import '../task-list/components/task-filters.component';
 @import '../task-list/components/task-header.component';
@@ -12,8 +11,7 @@
     @include adf-process-filters-theme($theme);
     @include adf-task-list-comment-list-theme($theme);
     @include adf-task-list-start-task-theme($theme);
-    @include adf-task-list-people-search-theme($theme);
-    @include adf-task-list-people-theme($theme);
+    @include adf-people-module-theme($theme);
     @include adf-task-list-filters-task-theme($theme);
     @include adf-task-list-header-theme($theme);
     @include adf-process-attachment-list-theme($theme);
